@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "^13.3.4",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
-    "@angular/material": "^13.3.4",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/animations": "~14.0.1",
+    "@angular/cdk": "^14.0.1",
+    "@angular/common": "~14.0.1",
+    "@angular/compiler": "~14.0.1",
+    "@angular/core": "~14.0.1",
+    "@angular/forms": "~14.0.1",
+    "@angular/material": "^14.0.1",
+    "@angular/platform-browser": "~14.0.1",
+    "@angular/platform-browser-dynamic": "~14.0.1",
+    "@angular/router": "~14.0.1",
     "json-server": "^0.17.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
@@ -29,7 +29,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.2",
     "@angular/cli": "~13.3.2",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular/compiler-cli": "~14.0.1",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
